requirements.txt
pandas
chardet
numpy
python-dateutil
CurrencyConverter
scikit-learn
imblearn
seaborn
dvc
-e .